+++ src/main/java/step3/domain/Frames.java
@@ -2,7 +2,6 @@ package step3.domain;
 
 import java.util.ArrayList;
 import java.util.List;
-import step3.NormalFrame;
 
 public class Frames {
     List<NormalFrame> frames;
