+++ src/main/java/step3/domain/NormalFrame.java
@@ -1,18 +1,32 @@
-package step3;
+package step3.domain;
 
-import step3.state.Ready;
-import step3.state.State;
+import step3.domain.state.Ready;
+import step3.domain.state.State;
 
-public class FinalFrame implements Frame {
+public class NormalFrame implements Frame {
+    private Frame next;
     private State state;
+    private int no;
 
-    public FinalFrame() {
+    public NormalFrame(int frameNum) {
         this.state = new Ready();
+        this.no = frameNum;
     }
 
     public Frame bowl(int fallenPins) {
         state = state.bowl(fallenPins);
+        if (state.isFinish()) {
+            next = createFrame();
+            return next;
+        }
         return this;
+    }
+
+    private Frame createFrame() {
+        if (no + 1 == 10) {
+            return new FinalFrame();
+        }
+        return new NormalFrame(no + 1);
     }
 
     public Score getScore() {
@@ -21,7 +35,7 @@ public class FinalFrame implements Frame {
             return score;
         }
 
-        return score;
+        return next.calculateAdditionalScore(score);
     }
 
     public Score calculateAdditionalScore(Score beforeScore) {
@@ -29,7 +43,7 @@ public class FinalFrame implements Frame {
         if (score.canCalculateScore()) {
             return score;
         }
-        return score;
+        return next.calculateAdditionalScore(score);
     }
 
     public boolean isGameEnd() {
@@ -40,7 +54,6 @@ public class FinalFrame implements Frame {
         return state;
     }
 
-    @Override
     public boolean isFinish() {
         return state.isFinish();
     }
