+++ src/main/java/step3/domain/Pins.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 import java.util.Objects;
 import step3.exceptions.PinNumberExecption;
