+++ src/main/java/step3/domain/state/FirstBowl.java
@@ -1,8 +1,8 @@
-package step3.state;
+package step3.domain.state;
 
 import java.util.Objects;
-import step3.Pins;
-import step3.Score;
+import step3.domain.Pins;
+import step3.domain.Score;
 import step3.exceptions.CanNotThrowBallException;
 
 public class FirstBowl implements State {
