+++ src/main/java/step3/domain/state/Ready.java
@@ -1,6 +1,6 @@
-package step3.state;
+package step3.domain.state;
 
-import step3.Score;
+import step3.domain.Score;
 import step3.exceptions.CannotCalculateExceptions;
 
 public class Ready implements State {
