+++ src/test/java/step3/PinsTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Pins;
 import step3.exceptions.PinNumberExecption;
 
 public class PinsTest {
