+++ src/test/java/step3/ScoreTest.java
@@ -2,9 +2,9 @@ package step3;
 
 import static org.assertj.core.api.Assertions.*;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Score;
 
 public class ScoreTest {
     @DisplayName("bowl 메서드 기능 확인 테스트")
